more floats

+++ Mobile Buy SDK/Mobile Buy SDK/Product View/Variant Selection/BUYOptionSelectionNavigationController.m
@@ -94,8 +94,8 @@
 		BUYOptionSelectionNavigationController *optionSelectionNavigationController = (BUYOptionSelectionNavigationController *)[transitionContext viewControllerForKey:UITransitionContextFromViewControllerKey];
 		if (optionSelectionNavigationController.dismissWithCancelAnimation) {
 			frame.origin.y += 150;
-			int angle = (int)arc4random_uniform(20) - 10;
-			transform = CGAffineTransformMakeRotation(((angle) / 180.0 * M_PI));
+			float angle = (int)arc4random_uniform(20.0f) - 10.0f;
+			transform = CGAffineTransformMakeRotation(((angle) / 180.0f * M_PI));
 		} else {
 			transform = CGAffineTransformMakeScale(2, 2);
 		}
